@@ -26,8 +26,8 @@
     "pg": "8.5.1",
     "pg-connection-string": "2.4.0",
     "reflect-metadata": "0.1.13",
-    "sequelize": "5.22.5",
-    "sequelize-typescript": "1.1.0",
+    "sequelize": "6.21.2",
+    "sequelize-typescript": "2.1.3",
     "source-map-support": "0.5.19"
   },
   "devDependencies": {
@@ -37,8 +37,8 @@
     "@types/node": "14.0.23",
     "@types/nodemailer": "6.4.0",
     "@types/validator": "13.1.0",
-    "tsc-watch": "4.2.9",
-    "typescript": "3.9.7"
+    "tsc-watch": "4.6.2",
+    "typescript": "4.7.4"
   },
   "_moduleAliases": {
     "@shared": "../../shared/build"
