@@ -1,4 +1,4 @@
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
 import styles from './Error.module.scss'
 
@@ -6,7 +6,7 @@ import heroImg from 'assets/img/not-found.png'
 import { PrimaryButton, TextButton } from 'components/common'
 
 const Error = () => {
-  const history = useHistory()
+  const navigate = useNavigate()
 
   return (
     <div className={styles.rootContainer}>
@@ -16,12 +16,12 @@ const Error = () => {
         The link you followed may be broken, or the page may not exist.
       </p>
       <span className={styles.actions}>
-        <PrimaryButton className={styles.action} onClick={history.goBack}>
+        <PrimaryButton className={styles.action} onClick={() => navigate(-1)}>
           Go back to previous page
         </PrimaryButton>
         <TextButton
           className={styles.action}
-          onClick={() => history.push('/campaigns')}
+          onClick={() => navigate('/campaigns')}
         >
           Go to dashboard
         </TextButton>
