@@ -7,7 +7,7 @@ import { I18nProvider } from '@lingui/react'
 import * as Sentry from '@sentry/browser'
 
 import { StrictMode } from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 
 // Locales
@@ -25,15 +25,16 @@ Sentry.init({
   environment: SENTRY_ENVIRONMENT,
 })
 
-ReactDOM.render(
-  <StrictMode>
-    <BrowserRouter>
+const container = document.getElementById('root') as HTMLElement
+const root = createRoot(container)
+root.render(
+  <BrowserRouter>
+    <StrictMode>
       <I18nProvider i18n={i18n}>
         <AuthContextProvider>
           <App />
         </AuthContextProvider>
       </I18nProvider>
-    </BrowserRouter>
-  </StrictMode>,
-  document.getElementById('root')
+    </StrictMode>
+  </BrowserRouter>
 )
