@@ -2,6 +2,7 @@
 
 module.exports = {
   up: async (queryInterface, _) => {
+    // NB superseded by 20221012051800-add-api-key-hash-column-to-user.js
     await queryInterface.createFunction(
       'format_user_on_insert', // actually a misnomer, since activates on UPDATE too, but stick to existing naming
       [],
@@ -16,12 +17,14 @@ module.exports = {
       { force: true }
     )
     await queryInterface.sequelize.query(
-      "CREATE TRIGGER format_user_trigger BEFORE INSERT OR UPDATE ON users FOR EACH ROW EXECUTE PROCEDURE format_user_on_insert();"
+      'CREATE TRIGGER format_user_trigger BEFORE INSERT OR UPDATE ON users FOR EACH ROW EXECUTE PROCEDURE format_user_on_insert();'
     )
   },
 
   down: async (queryInterface, _) => {
-    await queryInterface.sequelize.query("DROP TRIGGER format_user_trigger ON users;")
+    await queryInterface.sequelize.query(
+      'DROP TRIGGER format_user_trigger ON users;'
+    )
     await queryInterface.dropFunction('format_user_on_insert', [])
   },
 }
