@@ -44,18 +44,18 @@ export class EmailMessageTransactional extends Model<EmailMessageTransactional> 
   userId: string
 
   @Column({ type: DataType.STRING, allowNull: true })
-  from: string
+  from: string | null
 
   @Column({ type: DataType.STRING, allowNull: false })
   recipient: string
 
-  @Column({ type: DataType.JSON, allowNull: false })
+  @Column({ type: DataType.JSONB, allowNull: false })
   params: Record<string, string>
 
   @Column({ type: DataType.STRING, allowNull: true })
   messageId: string | null
 
-  @Column({ type: DataType.ARRAY(DataType.JSON), allowNull: true })
+  @Column({ type: DataType.ARRAY(DataType.JSONB), allowNull: true })
   attachmentsMetadata: AttachmentsMetadata | null
 
   @Column({
@@ -64,20 +64,13 @@ export class EmailMessageTransactional extends Model<EmailMessageTransactional> 
   })
   status: TransactionalEmailMessageStatus
 
-  /*
-   * see https://www.notion.so/opengov/23-Sep-2022-RFC-for-Saving-Emails-Sent-via-API-e4d29e9488004bcbb430ab15c9ef589f#bae59004a81846618552922e3751e012 for detailed discussion on errorCode and errorSubType
-   * */
-
-  // will only handle the non-callback error codes for now
   @Column({ type: DataType.STRING, allowNull: true })
   errorCode: string | null
 
-  // this is the error type returned by AWS SES; to be handled when callback is supported
-  // @Column({ type: DataType.STRING, allowNull: true })
-  // errorSubType: string | null
+  @Column({ type: DataType.STRING, allowNull: true })
+  errorSubType: string | null
 
   /*
-   * TODO: capture key timestamps in email's lifecycle and align with email_messages table
    * https://www.notion.so/opengov/Support-callbacks-in-transactional-emails-cade9647b2264a28a9a4d7eca301846a
    * */
   // this is equivalent to `sent_at` from `email_messages` table
