@@ -1,6 +1,6 @@
 import userEvent from '@testing-library/user-event'
 
-import { Route } from 'react-router-dom'
+import { Route, Routes } from 'react-router-dom'
 
 import EmailRecipients from '../EmailRecipients'
 
@@ -55,15 +55,20 @@ function renderRecipients() {
   const setActiveStep = jest.fn()
 
   render(
-    <Route path="/campaigns/:id">
-      <CampaignContextProvider
-        initialCampaign={new EmailCampaign({ ...TEST_EMAIL_CAMPAIGN })}
-      >
-        <FinishLaterModalContextProvider>
-          <EmailRecipients setActiveStep={setActiveStep} />
-        </FinishLaterModalContextProvider>
-      </CampaignContextProvider>
-    </Route>,
+    <Routes>
+      <Route
+        path="/campaigns/:id"
+        element={
+          <CampaignContextProvider
+            initialCampaign={new EmailCampaign({ ...TEST_EMAIL_CAMPAIGN })}
+          >
+            <FinishLaterModalContextProvider>
+              <EmailRecipients setActiveStep={setActiveStep} />
+            </FinishLaterModalContextProvider>
+          </CampaignContextProvider>
+        }
+      />
+    </Routes>,
     {
       router: { initialIndex: 0, initialEntries: ['/campaigns/1'] },
     }
