@@ -101,7 +101,10 @@ const SMSRecipients = ({
     const setSelectedList = async () => {
       try {
         if (selectedListId) {
-          await selectList({ campaignId: +campaignId, listId: selectedListId })
+          await selectList({
+            campaignId: +(campaignId as string),
+            listId: selectedListId,
+          })
           setIsCsvProcessing(true)
         }
       } catch (e) {
@@ -163,7 +166,7 @@ const SMSRecipients = ({
 
   // If shouldSaveList is modified, send info to backend
   useEffect(() => {
-    void setCampaignToSaveList(+campaignId, shouldSaveList)
+    void setCampaignToSaveList(campaignId as string, shouldSaveList)
   }, [campaignId, shouldSaveList])
 
   // Handle file upload
