@@ -1,6 +1,6 @@
 import userEvent from '@testing-library/user-event'
 
-import { Route } from 'react-router'
+import { Route, Routes } from 'react-router-dom'
 
 import TelegramCredentials from '../TelegramCredentials'
 
@@ -49,15 +49,22 @@ function renderTelegramCredentials() {
   const setActiveStep = jest.fn()
 
   render(
-    <Route path="/campaigns/:id">
-      <CampaignContextProvider
-        initialCampaign={new TelegramCampaign({ ...TEST_TELEGRAM_CAMPAIGN })}
-      >
-        <FinishLaterModalContextProvider>
-          <TelegramCredentials setActiveStep={setActiveStep} />
-        </FinishLaterModalContextProvider>
-      </CampaignContextProvider>
-    </Route>,
+    <Routes>
+      <Route
+        path="/campaigns/:id/*"
+        element={
+          <CampaignContextProvider
+            initialCampaign={
+              new TelegramCampaign({ ...TEST_TELEGRAM_CAMPAIGN })
+            }
+          >
+            <FinishLaterModalContextProvider>
+              <TelegramCredentials setActiveStep={setActiveStep} />
+            </FinishLaterModalContextProvider>
+          </CampaignContextProvider>
+        }
+      />
+    </Routes>,
     {
       router: {
         initialIndex: 0,
